@@ -5,39 +5,45 @@
     "dependencies": {
         "@netlify/ipx": "^1.3.3",
         "@netlify/plugin-nextjs": "^4.31.0",
-        "@next/bundle-analyzer": "^13.1.6",
+        "@next/bundle-analyzer": "^13.2.4",
         "@react-icons/all-files": "^4.1.0",
-        "axios": "^1.2.6",
-        "bip32": "^3.1.0",
+        "axios": "^1.3.4",
+        "bip32": "^4.0.0",
         "bitcoin-address-validation": "^2.2.1",
         "bitcoinjs-lib": "^6.1.0",
-        "bootstrap": "^5.1.3",
+        "bootstrap": "^5.2.3",
         "buffer": "^6.0.3",
         "clsx": "^1.2.1",
         "copy-webpack-plugin": "^11.0.0",
         "ecpair": "^2.1.0",
-        "framer-motion": "^9.0.3",
-        "mdb-react-ui-kit": "^4.0.0",
+        "env-var": "^7.3.0",
+        "framer-motion": "^10.2.4",
+        "lodash.difference": "^4.5.0",
+        "lodash.differenceby": "^4.8.0",
+        "lodash.isequal": "^4.5.0",
+        "lodash.uniqby": "^4.7.0",
+        "mdb-react-ui-kit": "^5.1.0",
         "netlify-plugin-ipx": "0.0.1",
-        "next": "^13.1.6",
-        "next-themes": "^0.1.1",
-        "node-polyfill-webpack-plugin": "^2.0.0",
-        "nostr-tools": "^1.6.0",
+        "next": "^13.2.4",
+        "next-themes": "^0.2.1",
+        "node-polyfill-webpack-plugin": "^2.0.1",
+        "nostr-tools": "^1.7.4",
         "prop-types": "^15.8.1",
         "react": "^18.2.0",
-        "react-bootstrap": "^2.4.0",
-        "react-countdown": "^2.3.2",
+        "react-bootstrap": "^2.7.2",
+        "react-countdown": "^2.3.5",
         "react-dom": "^18.2.0",
-        "react-icons": "^4.7.1",
+        "react-icons": "^4.8.0",
         "react-loading-icons": "^1.1.0",
         "react-rotating-text": "^1.4.1",
         "react-scripts": "^5.0.1",
         "react-scroll": "^1.8.9",
         "react-slick": "^0.29.0",
-        "react-slide-fade-in": "^1.0.7",
+        "react-slide-fade-in": "^1.0.9",
         "react-social-icons": "^5.15.0",
         "react-toastify": "^9.1.1",
         "react-useinterval": "^1.0.2",
+        "rxjs": "^7.8.0",
         "sal.js": "^0.8.5",
         "sharp": "^0.31.3",
         "tiny-secp256k1": "^2.2.1"
@@ -69,21 +75,21 @@
     "author": "",
     "license": "ISC",
     "devDependencies": {
-        "@babel/core": "^7.16.0",
-        "@babel/eslint-parser": "^7.16.3",
+        "@babel/core": "^7.21.0",
+        "@babel/eslint-parser": "^7.19.1",
         "@babel/preset-react": "^7.18.6",
-        "@next/eslint-plugin-next": "^13.0.6",
-        "eslint": "8.4.1",
-        "eslint-config-airbnb": "^19.0.2",
-        "eslint-config-prettier": "^8.3.0",
+        "@next/eslint-plugin-next": "^13.2.4",
+        "eslint": "8.35.0",
+        "eslint-config-airbnb": "^19.0.4",
+        "eslint-config-prettier": "^8.7.0",
         "eslint-import-resolver-alias": "^1.1.2",
-        "eslint-plugin-import": "^2.25.3",
-        "eslint-plugin-jsx-a11y": "^6.5.1",
-        "eslint-plugin-prettier": "^4.0.0",
-        "eslint-plugin-react": "^7.27.1",
-        "eslint-plugin-react-hooks": "^4.3.0",
-        "prettier": "^2.5.1",
-        "rimraf": "^3.0.2",
-        "sass": "^1.58.1"
+        "eslint-plugin-import": "^2.27.5",
+        "eslint-plugin-jsx-a11y": "^6.7.1",
+        "eslint-plugin-prettier": "^4.2.1",
+        "eslint-plugin-react": "^7.32.2",
+        "eslint-plugin-react-hooks": "^4.6.0",
+        "prettier": "^2.8.4",
+        "rimraf": "^4.4.0",
+        "sass": "^1.58.3"
     }
 }
