@@ -1,17 +1,17 @@
 /* eslint-disable react/forbid-prop-types */
 /* eslint-disable no-restricted-syntax */
 /* eslint-disable no-extra-boolean-cast */
-import { useContext, useState, useEffect } from "react";
+import { useContext, useState, useEffect, useCallback, useRef } from "react";
 import PropTypes from "prop-types";
 import clsx from "clsx";
 import SectionTitle from "@components/section-title";
 import OrdinalCard from "@components/ordinal-card";
 import { deepClone } from "@utils/methods";
-import OpenOrdex from "@utils/openOrdexV3";
-import SessionStorage, { SessionsStorageKeys } from "@services/session-storage";
 import WalletContext from "@context/wallet-context";
-// Use this to fetch data from an API service
-const axios = require("axios");
+import { nostrPool } from "@services/nostr-relay";
+import { MAX_ONSALE } from "@lib/constants.config";
+import { Subject } from "rxjs";
+import { scan } from "rxjs/operators";
 
 const collectionAuthor = [
     {
@@ -26,50 +26,83 @@ const collectionAuthor = [
 const OnSaleOrdinalsArea = ({ className, space, onConnectHandler, onSale }) => {
     const { nostrAddress, isExperimental } = useContext(WalletContext);
     const [openOrders, setOpenOrders] = useState([]);
-    const [isLoadingOpenOrders, setIsLoadingOpenOrders] = useState(true);
+    const [isLoadingOpenOrders, setIsLoadingOpenOrders] = useState(false); // it is necessary?
+    const addOpenOrder$ = useRef(new Subject());
+    const addSubscriptionRef = useRef(null);
+    const orderSubscriptionRef = useRef(null);
+    const [isWindowFocused, setIsWindowFocused] = useState(true);
+
+    const addNewOpenOrder = (order) => {
+        addOpenOrder$.current.next(order);
+    };
+
+    const formatOrder = useCallback((inscription) => {
+        const inscriptionData = Object.assign(
+            {},
+            ...inscription.tags
+                .map(([tagId, value]) => ({
+                    [tagId]: value,
+                }))
+                .map((o) => o)
+        );
+        const forSaleInscription = deepClone({
+            inscriptionTags: inscriptionData,
+            ...inscription,
+        });
+        return forSaleInscription;
+    }, []);
 
     useEffect(() => {
-        console.log("OnSaleOrdinalsArea useEffect");
-        // safe on session storage for faster loads
-        window.addEventListener("message", (event) => {
-            console.debug(event);
-        });
-        const load = async () => {
-            setIsLoadingOpenOrders(true);
-            // load from cache before updating
-            const sessionOrders = SessionStorage.get(SessionsStorageKeys.INSCRIPTIONS_ON_SALE);
-            if (sessionOrders) {
-                setOpenOrders(sessionOrders);
-            }
-            const openOrderx = await OpenOrdex.init();
-            const orders = await openOrderx.getLatestOrders(25);
-
-            const forSaleInscriptions = [];
-            for (const inscription of orders) {
-                let inscriptionData = inscription.tags
-                    // .filter(([t, v]) => t === "i" && v)
-                    .map(([tagId, value]) => ({
-                        [tagId]: value,
-                    }));
-                // Convert array into object of key tagId
-                inscriptionData = Object.assign({}, ...inscriptionData.map((o) => o));
-
-                const i = deepClone({
-                    inscriptionTags: inscriptionData,
-                    ...inscription,
-                });
-                // console.log(i);
-
-                forSaleInscriptions.push(i);
-            }
-
-            SessionStorage.set(SessionsStorageKeys.INSCRIPTIONS_ON_SALE, forSaleInscriptions);
-
-            setOpenOrders(forSaleInscriptions);
-
-            setIsLoadingOpenOrders(false);
+        if (isWindowFocused) {
+            addSubscriptionRef.current = addOpenOrder$.current
+                .pipe(
+                    scan((acc, curr) => {
+                        // We can add only unique ordinals by id or inscriptionId
+                        // I keep if for test purposes
+                        if (acc.find((order) => order.id === curr.id)) {
+                            return acc;
+                        }
+                        // We sort by created_at DESC and limit list
+                        return [...acc, curr].sort((a, b) => b.created_at - a.created_at).slice(0, MAX_ONSALE);
+                    }, openOrders)
+                )
+                .subscribe(setOpenOrders);
+            orderSubscriptionRef.current = nostrPool.subscribeOrders({ limit: MAX_ONSALE }).subscribe((order) => {
+                addNewOpenOrder(formatOrder(order));
+            });
+        }
+        return () => {
+            try {
+                orderSubscriptionRef?.current?.unsubscribe();
+                addSubscriptionRef?.current?.unsubscribe();
+                // eslint-disable-next-line no-empty
+            } catch (err) {}
         };
-        load();
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [isWindowFocused]);
+
+    useEffect(() => {
+        const handleWindowBlur = () => {
+            setIsWindowFocused(false);
+        };
+
+        const handleWindowFocus = () => {
+            setIsWindowFocused(true);
+        };
+
+        const handleVisibilityChange = () => {
+            setIsWindowFocused(!document.hidden);
+        };
+
+        window.addEventListener("blur", handleWindowBlur);
+        window.addEventListener("focus", handleWindowFocus);
+        document.addEventListener("visibilitychange", handleVisibilityChange);
+
+        return () => {
+            window.removeEventListener("blur", handleWindowBlur);
+            window.removeEventListener("focus", handleWindowFocus);
+            document.removeEventListener("visibilitychange", handleVisibilityChange);
+        };
     }, []);
 
     return (
@@ -80,8 +113,10 @@ const OnSaleOrdinalsArea = ({ className, space, onConnectHandler, onSale }) => {
                         <SectionTitle
                             className="mb--0 with-loading"
                             isLoading={isLoadingOpenOrders}
-                            {...{ title: "On sale" }}
+                            {...{ title: "On Sale" }}
+                            subtitle="Buy & sell comming soon"
                         />
+                        <span>Buy and sell comming soon</span>
                         {!Boolean(nostrAddress) && isExperimental && (
                             <span>
                                 <button type="button" className="btn-transparent" onClick={onConnectHandler}>
@@ -92,34 +127,31 @@ const OnSaleOrdinalsArea = ({ className, space, onConnectHandler, onSale }) => {
                         )}
                     </div>
                 </div>
-
-                {!!openOrders.length && (
-                    <div className="row g-5">
-                        {openOrders.length > 0 ? (
-                            <>
-                                {openOrders.map((utxo) => (
-                                    <div key={utxo.id} className="col-5 col-lg-4 col-md-6 col-sm-6 col-12">
-                                        <OrdinalCard
-                                            overlay
-                                            price={{
-                                                amount: utxo.value.toLocaleString("en-US"),
-                                                currency: "Sats",
-                                            }}
-                                            type="buy"
-                                            confirmed
-                                            date={utxo.created_at}
-                                            authors={collectionAuthor}
-                                            utxo={utxo}
-                                            onSale={onSale}
-                                        />
-                                    </div>
-                                ))}
-                            </>
-                        ) : (
-                            <div>There are no inscriptions for sale yet..</div>
-                        )}
-                    </div>
-                )}
+                <div className="row g-5">
+                    {openOrders.length > 0 ? (
+                        <>
+                            {openOrders.map((utxo) => (
+                                <div key={utxo.id} className="col-5 col-lg-4 col-md-6 col-sm-6 col-12">
+                                    <OrdinalCard
+                                        overlay
+                                        price={{
+                                            amount: utxo.value.toLocaleString("en-US"),
+                                            currency: "Sats",
+                                        }}
+                                        type="buy"
+                                        confirmed
+                                        date={utxo.created_at}
+                                        authors={collectionAuthor}
+                                        utxo={utxo}
+                                        onSale={onSale}
+                                    />
+                                </div>
+                            ))}
+                        </>
+                    ) : (
+                        <div>There are no inscriptions for sale yet..</div>
+                    )}
+                </div>
             </div>
         </div>
     );
